Fix: Change location of log out button functionality.
